+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 5.1.3
+
+- Introduces `ModelFill`, an extension of `Fill`. When set on a `Style`, the color
+  will also be applied to models which are added via the `olcs_modelUrl` property.
+- `getStyleOrDefaultStyle` no longer mutates default style when passing in a vector default style.
+
 ### 5.1.2
 
 - Fixes a bug where anchors from flight GeoJSON files were not added correctly.